extract 'getShopifyToken' to utils
