+++ src/helpers/utils.js
@@ -1,5 +1,6 @@
 const util = require('util');
 const jwt = require('jsonwebtoken');
+const ShopifyToken = require('shopify-token');
 
 const validationError = (result) => {
   return 'There have been validation errors: ' + util.inspect(result.array());
@@ -20,7 +21,15 @@ const getJWTToken = (shop) => {
   return jwt.sign(payload, process.env.SHOPIFY_APP_SECRET, { expiresIn: '3h' });
 }
 
+const getShopifyToken = () => {
+  return new ShopifyToken({
+    sharedSecret: process.env.SHOPIFY_APP_SECRET,
+    redirectUri: `${process.env.BASE_URL}/auth/redirect/uri`,
+    apiKey: process.env.SHOPIFY_API_KEY
+  });
+};
+
 
 module.exports = {
-  validationError, groupBy, getJWTToken
+  validationError, groupBy, getJWTToken, getShopifyToken
 };
